refactor theme component into using useContext to pass down theme through provider
