+++ src/App.jsx
@@ -1,14 +1,15 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import React, { createContext, useContext, useState } from "react";
+import React from "react";
+import FunctionContextComponent from "./Functions/FunctionContextComponent.jsx";
 import Homepage from "./pages/Homepage/Homepage";
 import Details from "./pages/Details/Details";
 import Cart from "./pages/Cart/Cart";
+import { ThemeProvider } from "./Context/ThemeContext.jsx";
 
-const shoppingCartContext = createContext();
 function App() {
-  const [shoppingCart, setShoppingCart] = useState([]); //initialize shopping cart
   return (
-    <shoppingCartContext.Provider value={{ shoppingCart, setShoppingCart }}>
+    <ThemeProvider>
+      <FunctionContextComponent />
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Homepage />} />
@@ -16,7 +17,7 @@ function App() {
           <Route path="/cart" element={<Cart />} />
         </Routes>
       </BrowserRouter>
-    </shoppingCartContext.Provider>
+    </ThemeProvider>
   );
 }
