+++ src/Context/ThemeContext.jsx
@@ -1,0 +1,35 @@
+import React, { useEffect, useState, useContext } from "react";
+
+const ThemeContext = React.createContext();
+const ThemeUpdateContext = React.createContext();
+
+export function useTheme() {
+  return useContext(ThemeContext);
+}
+
+export function useThemeUpdate() {
+  return useContext(ThemeUpdateContext);
+}
+
+export function ThemeProvider({ children }) {
+  const [theme, setTheme] = useState(() => {
+    return localStorage.getItem("theme") || "light";
+  });
+
+  useEffect(() => {
+    document.documentElement.setAttribute("data-theme", theme);
+    localStorage.setItem("theme", theme);
+  }, [theme]);
+
+  const toggleTheme = () => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  };
+
+  return (
+    <ThemeContext.Provider value={theme}>
+      <ThemeUpdateContext.Provider value={toggleTheme}>
+        {children}
+      </ThemeUpdateContext.Provider>
+    </ThemeContext.Provider>
+  );
+}
