+++ src/styles/_variables.scss
@@ -9,8 +9,6 @@ $bg-beige: #f7f2ed;
 $light-beige: #faf9f6;
 $icons: #635443;
 
-//with light and dark mode in mind
-
 //default
 $light-primary-text: #2f2f2f;
 $light-tertiary-text: #ffffff;
